It works!!!
Added code to convert the jar to
executable. It should work as planned,
though slight tweaks might be necessary. As of now,
an entire directory also needs to be cut and paste
in the plugins directory, and I need to refine the
display to be more userfriendly.

I added a few import statements specifically
for ImageJ, changed the class declaration, changed the
main method to be called "run," and then a few statements
to display the output in a separate ImageJ text window.

This did not add much code but finding the necessary lines
was a beast.